+++ .gitignore
@@ -1,9 +1,8 @@
-*.js
 !jest.config.js
-*.d.ts
 node_modules
 .idea
 .eslintcache
+dist
 
 # CDK asset staging directory
 .cdk.staging
